requirements.txt
flask==2.3.3
flask-restx==1.1.0
python-dotenv==1.0.0
